backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# Environment variables
python-dotenv==1.0.0

# AI/LLM Integration
google-generativeai==0.8.3

# Date/Time
python-dateutil==2.8.2

# HTTP Client
httpx==0.26.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.13
